growex_bot/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.20.0.post0",
    "python-dotenv>=1.1.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "schedule>=1.2.0",
    "cryptography>=41.0.0",
]
